+++ TinkoffASDKCore/TinkoffASDKCore/PaymentInitRequest.swift
@@ -111,6 +111,7 @@ public struct PaymentInitData: Codable {
         try container.encode(amount, forKey: .amount)
         try container.encode(orderId, forKey: .orderId)
         try container.encodeIfPresent(customerKey, forKey: .customerKey)
+        try container.encodeIfPresent(payType?.rawValue, forKey: .payType)
         if description != nil { try? container.encode(description, forKey: .description) }
         if redirectDueDate != nil { try? container.encode(redirectDueDate, forKey: .redirectDueDate) }
         if let value = savingAsParentPayment, value == true { try container.encode("Y", forKey: .savingAsParentPayment) }
@@ -123,21 +124,24 @@ public struct PaymentInitData: Codable {
     public init(amount: Int64,
                 orderId: String,
                 customerKey: String?,
-                redirectDueDate: Date? = nil)
+                redirectDueDate: Date? = nil,
+                payType: PayType? = nil)
     {
         self.amount = amount
         self.orderId = orderId
         self.customerKey = customerKey
         self.redirectDueDate = redirectDueDate
+        self.payType = payType
     }
 
     public init(amount: NSDecimalNumber,
                 orderId: String,
                 customerKey: String?,
-                redirectDueDate: Date? = nil)
+                redirectDueDate: Date? = nil,
+                payType: PayType? = nil)
     {
         let int64Amount = Int64(amount.doubleValue * 100)
-        self.init(amount: int64Amount, orderId: orderId, customerKey: customerKey, redirectDueDate: redirectDueDate)
+        self.init(amount: int64Amount, orderId: orderId, customerKey: customerKey, redirectDueDate: redirectDueDate, payType: payType)
     }
 }