+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 ## [Unreleased]
 
+## [2.6.0] - 2021-12-19
+
+### Added
+* add sending the PayType parameter to the Init method 
+* add completion handler to urlSBPPaymentViewController AcquiringUISDK method
+
 ## [2.5.1] - 2021-12-06
 
 ### Updated